+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - history of changes: see https://github.com/delving/hub3/compare/v0.1.10...master
 
+### Fixes
+
+- Delete DataSet does not work when when no v1 index is created [[GH-25]](https://github.com/delving/hub3/pull/25)
+
 ## v0.1.10 (2020-07-20)
 
 - history of changes: see https://github.com/delving/hub3/compare/v0.1.9...v0.1.10